refactor(ui-codemods): finish breaking codemods for v8 upgrade
