+++ packages/ui-codemods/lib/utils/updateV7Lists.ts
@@ -30,7 +30,7 @@ import {
   findImport,
   printWarning,
   renameElements
-} from '../helpers/v7PropsUpdateHelpers'
+} from '../helpers/codemodHelpers'
 
 /**
  * Does the following update on a <List>:
